# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iGO - API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/client/cart": {
            "get": {
                "tags": [
                    "Client: Cart"
                ],
                "summary": "Show products in the cart",
                "description": "Display a list of products added to the cart",
                "operationId": "showCart",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Client: Cart"
                ],
                "summary": "Add item to the cart",
                "description": "Create/Update a product to/in the cart - Expect to recieve a product id and a quantity - If the product is already in the cart, updates the quantity",
                "operationId": "addToCart",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": "4"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "side": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "sauce": {
                                        "type": "integer",
                                        "example": "6"
                                    },
                                    "extras": {
                                        "type": "string",
                                        "example": "[{'extra_id':'3','extra_quantity':'1'}]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/client/cart/{id}": {
            "delete": {
                "tags": [
                    "Client: Cart"
                ],
                "summary": "Remove item from the cart",
                "description": "Remove a specified product from the cart - Expect to recieve a valid product id",
                "operationId": "removeFromCart",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/client/profile": {
            "get": {
                "tags": [
                    "Client: Profile & Favorite partners"
                ],
                "summary": "Get client profile",
                "description": "Get data from logged in client",
                "operationId": "getClientProfile",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Client: Profile & Favorite partners"
                ],
                "summary": "Set client profile",
                "description": "Update client profile data",
                "operationId": "setClientProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mussum Ipsum"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "09/07/1988"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "mussum@igo.pt"
                                    },
                                    "mobile_phone_number": {
                                        "type": "integer",
                                        "example": "978 645 312"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "iGO@123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/client/favorite/{id}": {
            "post": {
                "tags": [
                    "Client: Profile & Favorite partners"
                ],
                "summary": "Set favorite partners",
                "description": "Set/Unset a partner as favorite - Expect to receive a valid partner id",
                "operationId": "setFavoritePartners",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Partner id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}